Added code for pull server removal

+++ DSCResources/MSFT_xDSCWebService/PSWSIISEndpoint.psm1
@@ -4,6 +4,10 @@
 #Copyright (c) Microsoft Corporation, 2014
 #
 
+# name and description for the Firewall rules. Used in multiple locations
+$FireWallRuleDisplayName = "Desired State Configuration - Pull Server Port:{0}"
+$FireWallRuleDescription = "Inbound traffic for IIS site on Port:{0} for DSC pull server. Created by DSCWebService resource"
+
 # Validate supplied configuration to setup the PSWS Endpoint
 # Function checks for the existence of PSWS Schema files, IIS config
 # Also validate presence of IIS on the target machine
@@ -61,12 +65,19 @@ function Initialize-Endpoint
     Test-IISInstall
     
     $appPool = "PSWS"
+
     
     Write-Verbose "Delete the App Pool if it exists"
     Remove-AppPool -apppool $appPool
-    
+   
     Write-Verbose "Remove the site if it already exists"
     Update-Site -siteName $site -siteAction Remove
+
+    # check for existing binding, there should be no binding with the same port
+    if ((Get-WebBinding | where bindingInformation -eq "*:$($port):").count -gt 0)
+    {
+        throw "ERROR: Port $port is already used, please review existing sites and change the port to be used." 
+    }
     
     if ($removeSiteFiles)
     {
@@ -77,11 +88,7 @@ function Initialize-Endpoint
     }
     
     Copy-Files -path $path -cfgfile $cfgfile -svc $svc -mof $mof -dispatch $dispatch -asax $asax -dependentBinaries $dependentBinaries -language $language -dependentMUIFiles $dependentMUIFiles -psFiles $psFiles
-    
-    Update-AllSites Stop
-    Update-DefaultAppPool Stop
-    Update-DefaultAppPool Start
-    
+   
     New-IISWebSite -site $site -path $path -port $port -app $app -apppool $appPool -applicationPoolIdentityType $applicationPoolIdentityType -certificateThumbPrint $certificateThumbPrint
 }
 
@@ -90,9 +97,9 @@ function Initialize-Endpoint
 function Test-IISInstall
 {
         Write-Verbose "Checking IIS requirements"
-        $iisVersion = (Get-ItemProperty HKLM:\SOFTWARE\Microsoft\InetStp -ErrorAction silentlycontinue).MajorVersion + (Get-ItemProperty HKLM:\SOFTWARE\Microsoft\InetStp -ErrorAction silentlycontinue).MinorVersion
+        $iisVersion = (Get-ItemProperty HKLM:\SOFTWARE\Microsoft\InetStp -ErrorAction silentlycontinue).MajorVersion
         
-        if ($iisVersion -lt 7.0) 
+        if ($iisVersion -lt 7) 
         {
             throw "ERROR: IIS Version detected is $iisVersion , must be running higher than 7.0"            
         }        
@@ -149,14 +156,15 @@ function Update-Site
         $name = $site.Name
     }
     
-    if (Test-IISSiteExists $name)
-    {        
+    if (Test-IISSiteExists -siteName $name)
+    {
         switch ($siteAction) 
         { 
-            "Start"  {Start-Website -Name $name} 
-            "Stop"   {Stop-Website -Name $name -ErrorAction SilentlyContinue} 
-            "Remove" {Remove-Website -Name $name}
+            "Start"  {Start-Website -Name "$name"} 
+            "Stop"   {Stop-Website -Name "$name" -ErrorAction SilentlyContinue} 
+            "Remove" {Remove-Website -Name "$name"}
         }
+        Write-Verbose "p11"
     }
 }
 
@@ -166,33 +174,11 @@ function Update-Site
 function Remove-AppPool
 {
     param ($appPool)    
-    
-    Remove-WebAppPool -Name $appPool -ErrorAction SilentlyContinue
-}
 
-# Perform given action(start, stop, delete) on all IIS Sites
-#
-function Update-AllSites
-{
-    param ($action)    
-    
-    foreach ($site in Get-Website)
+    # without this tests we may get a breaking error here, despite SilentlyContinue
+    if (Test-Path "IIS:\AppPools\$appPool")
     {
-        Update-Site $site $action
-    }
-}
-
-# Perform given action(start, stop) on the default app pool
-#
-function Update-DefaultAppPool
-{
-    param ($action) 
-    
-    switch ($action) 
-    { 
-        "Start"  {Start-WebAppPool -Name "DefaultAppPool"} 
-        "Stop"   {Stop-WebAppPool -Name "DefaultAppPool"} 
-        "Remove" {Remove-WebAppPool -Name "DefaultAppPool"}
+        Remove-WebAppPool -Name $appPool -ErrorAction SilentlyContinue
     }
 }
 
@@ -265,7 +251,7 @@ function Copy-Files
     $binFolderPath = Join-Path $path "bin"
     $null = New-Item -path $binFolderPath  -itemType "directory" -Force
     Copy-Item $dependentBinaries $binFolderPath -Force
-    
+
     if ($language)
     {
         $muiPath = Join-Path $binFolderPath $language
@@ -276,7 +262,7 @@ function Copy-Files
         }
         Copy-Item $dependentMUIFiles $muiPath -Force
     }
-    
+
     foreach ($psFile in $psFiles)
     {
         if (!(Test-Path $psFile))
@@ -372,9 +358,14 @@ function New-FirewallRule
     
     Write-Verbose "Disable Inbound Firewall Notification"
     Set-NetFirewallProfile -Profile Domain,Public,Private –NotifyOnListen False
+
+    $ruleDisplayName = ($($FireWallRuleDisplayName) -f $firewallPort)
+
+    # remove all existing rules with that displayName
+    Get-NetFirewallRule | Where-Object DisplayName -eq $ruleDisplayName | Remove-NetFirewallRule
     
     Write-Verbose "Add Firewall Rule for port $firewallPort"    
-    $null = New-NetFirewallRule -DisplayName "Allow Port $firewallPort for PSWS" -Direction Inbound -LocalPort $firewallPort -Protocol TCP -Action Allow
+    $null = New-NetFirewallRule -DisplayName $ruleDisplayName -Description ($($FireWallRuleDescription) -f $firewallPort) -Direction Inbound -LocalPort $firewallPort -Protocol TCP -Action Allow
 }
 
 # Enable & Clear PSWS Operational/Analytic/Debug ETW Channels
@@ -404,7 +395,7 @@ function Enable-PSWSETW
    Creates a PSWS IIS Endpoint by consuming PSWS Schema and related dependent files
 .EXAMPLE
    New a PSWS Endpoint [@ http://Server:39689/PSWS_Win32Process] by consuming PSWS Schema Files and any dependent scripts/binaries
-   New-PSWSEndpoint -site Win32Process -path $env:HOMEDRIVE\inetpub\wwwroot\PSWS_Win32Process -cfgfile Win32Process.config -port 39689 -app Win32Process -svc PSWS.svc -mof Win32Process.mof -dispatch Win32Process.xml -dependentBinaries ConfigureProcess.ps1, Rbac.dll -psFiles Win32Process.psm1
+   New-PSWSEndpoint -site Win32Process -path $env:SystemDrive\inetpub\PSWS_Win32Process -cfgfile Win32Process.config -port 39689 -app Win32Process -svc PSWS.svc -mof Win32Process.mof -dispatch Win32Process.xml -dependentBinaries ConfigureProcess.ps1, Rbac.dll -psFiles Win32Process.psm1
 #>
 function New-PSWSEndpoint
 {
@@ -414,8 +405,8 @@ function New-PSWSEndpoint
         # Unique Name of the IIS Site        
         [String] $site = "PSWS",
         
-        # Physical path for the IIS Endpoint on the machine (under inetpub/wwwroot)        
-        [String] $path = "$env:HOMEDRIVE\inetpub\wwwroot\PSWS",
+        # Physical path for the IIS Endpoint on the machine (under inetpub)        
+        [String] $path = "$env:SystemDrive\inetpub\PSWS",
         
         # Web.config file        
         [String] $cfgfile = "web.config",
@@ -502,10 +493,65 @@ function New-PSWSEndpoint
     if ($EnablePSWSETW)
     {
         Enable-PSWSETW
-    }
-    
-    Update-AllSites start
-    
+    }       
+}
+
+<#
+.Synopsis
+   Removes a DSC WebServices IIS Endpoint
+.DESCRIPTION
+   Removes a PSWS IIS Endpoint
+.EXAMPLE
+   Remove the endpoint with the specified name
+   Remove-PSWSEndpoint -siteName PSDSCPullServer 
+#>
+function Remove-PSWSEndpoint
+{
+[CmdletBinding()]
+    param (        
+        # Unique Name of the IIS Site        
+            [String] $siteName
+        )
+                
+       # get the site to remove
+       $site = Get-Item -Path "IIS:\sites\$siteName"
+       # and the pool it is using
+       $pool = $site.applicationPool
+
+       # get the path so we can delete the files
+       $filePath = $site.PhysicalPath
+       # get the port number for the Firewall rule
+       $bindings = (Get-WebBinding -Name $siteName).bindingInformation
+       $port = [regex]::match($bindings,':(\d+):').Groups[1].Value     
+
+       # remove the actual site.
+       Remove-Website -Name $siteName
+       # there may be running requests, wait a little
+       # I had an issue where the files were still in use
+       # when I tried to delete them
+       Start-Sleep -Milliseconds 200  
+
+       # remove the files for the site
+       If (Test-Path $filePath)
+       {
+           Get-ChildItem $filePath -Recurse | Remove-Item -Recurse
+           Remove-Item $filePath
+       }
+
+       # find out whether any other site is using this pool
+       $filter = "/system.applicationHost/sites/site/application[@applicationPool='" + $pool + "']" 
+       $apps = (Get-WebConfigurationProperty -Filter $filter -PSPath "machine/webroot/apphost" -name path).ItemXPath 
+       if ($apps.count -eq 1)
+       {
+          # if we are the only site in the pool, remove the pool as well.
+          Remove-WebAppPool -Name $pool
+       }
+
+
+       # remove all rules with that name
+       $ruleName = ($($FireWallRuleDisplayName) -f $port)
+       Get-NetFirewallRule | Where-Object DisplayName -eq "$ruleName" | Remove-NetFirewallRule
+
 }
 
 <#
@@ -519,7 +565,7 @@ function Set-AppSettingsInWebconfig
 {
     param (
                 
-        # Physical path for the IIS Endpoint on the machine (possibly under inetpub/wwwroot)
+        # Physical path for the IIS Endpoint on the machine (possibly under inetpub)
         [parameter(Mandatory)]
         [ValidateNotNullOrEmpty()]
         [String] $path,
@@ -589,7 +635,7 @@ function Set-BindingRedirectSettingInWebConfig
 {
     param (
                 
-        # Physical path for the IIS Endpoint on the machine (possibly under inetpub/wwwroot)
+        # Physical path for the IIS Endpoint on the machine (possibly under inetpub)
         [parameter(Mandatory)]
         [ValidateNotNullOrEmpty()]
         [String] $path,
@@ -660,4 +706,4 @@ function Set-BindingRedirectSettingInWebConfig
     }
 }
 
-Export-ModuleMember -function New-PSWSEndpoint, Set-AppSettingsInWebconfig, Set-BindingRedirectSettingInWebConfig
+Export-ModuleMember -function New-PSWSEndpoint, Set-AppSettingsInWebconfig, Set-BindingRedirectSettingInWebConfig, Remove-PSWSEndpoint
